refactor: route NextAuth callbacks through auth DAL/DTO
Replace the inline /auth/google fetch in the JWT callback and the
standalone src/data/guest-user.ts helper with getGoogleAuthView /
getGuestAuthView. Removes src/data/ and the now-empty types/dto.ts,
and extends the ESLint DAL boundary to block direct imports of
lib/dal/auth from app/ and components/.

Extract ApiError and getBaseUrl into lib/dal/http.ts to break the
import cycle that would otherwise form via lib/auth -> lib/dto/auth
-> lib/dal/auth -> lib/dal/client -> lib/dal/session -> lib/auth.
