+++ src/lib/auth/index.ts
@@ -1,5 +1,5 @@
 import NextAuth from 'next-auth';
-import type { AuthResponse } from '@/types/dto';
+import { getGoogleAuthView } from '@/lib/dto/auth';
 import { appProviders } from './app-providers';
 import { getAndClearGuestTokenCookie } from './guest-migration';
 
@@ -29,44 +29,11 @@ export const { handlers, signIn, signOut, auth } = NextAuth({
           throw new Error('Google ID token is missing');
         }
 
-        const apiUrl = process.env.API_URL;
-
-        if (apiUrl === undefined) {
-          throw new Error('API_URL is not configured');
-        }
-
         const guestToken = await getAndClearGuestTokenCookie();
 
-        try {
-          const response = await fetch(`${apiUrl}/auth/google`, {
-            method: 'POST',
-            headers: {
-              Authorization: `Bearer ${googleIdToken}`,
-              ...(guestToken !== undefined
-                ? { 'Content-Type': 'application/json' }
-                : {}),
-            },
-            ...(guestToken !== undefined
-              ? { body: JSON.stringify({ guest_token: guestToken }) }
-              : {}),
-          });
-
-          if (!response.ok) {
-            console.error(
-              'API authentication failed:',
-              response.status,
-              response.statusText,
-            );
-            throw new Error(`Authentication failed: ${response.status}`);
-          }
-
-          const result = (await response.json()) as AuthResponse;
-          token.accessToken = result.token;
-          token.isGuest = false;
-        } catch (error) {
-          console.error('API connection error:', error);
-          throw error;
-        }
+        const auth = await getGoogleAuthView(googleIdToken, guestToken);
+        token.accessToken = auth.token;
+        token.isGuest = false;
       }
 
       return token;
